+++ slang/test/integration/main/json_multiple_files_test.dart
@@ -3,7 +3,7 @@ import 'package:slang/builder/decoder/json_decoder.dart';
 import 'package:slang/builder/generator_facade.dart';
 import 'package:slang/builder/model/build_config.dart';
 import 'package:slang/builder/model/i18n_locale.dart';
-import 'package:slang/builder/model/namespace_translation_map.dart';
+import 'package:slang/builder/model/translation_map.dart';
 import 'package:test/test.dart';
 
 import '../../util/build_config_utils.dart';
@@ -34,7 +34,7 @@ void main() {
         outputFormat: OutputFormat.multipleFiles,
       ),
       baseName: 'translations',
-      translationMap: NamespaceTranslationMap()
+      translationMap: TranslationMap()
         ..addTranslations(
           locale: I18nLocale.fromString('en'),
           translations: JsonDecoder().decode(enInput),