+++ slang/test/integration/main/no_flutter_test.dart
@@ -2,7 +2,7 @@ import 'package:slang/builder/builder/build_config_builder.dart';
 import 'package:slang/builder/decoder/json_decoder.dart';
 import 'package:slang/builder/generator_facade.dart';
 import 'package:slang/builder/model/i18n_locale.dart';
-import 'package:slang/builder/model/namespace_translation_map.dart';
+import 'package:slang/builder/model/translation_map.dart';
 import 'package:test/test.dart';
 
 import '../../util/build_config_utils.dart';
@@ -25,7 +25,7 @@ void main() {
         flutterIntegration: false,
       ),
       baseName: 'translations',
-      translationMap: NamespaceTranslationMap()
+      translationMap: TranslationMap()
         ..addTranslations(
           locale: I18nLocale.fromString('en'),
           translations: JsonDecoder().decode(input),