+++ slang_build_runner/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 2.2.0
+
+- Read directory locale
+
 ## 2.1.0
 
 - Move `build.yaml`